use only index for post resource
The only use case for rendering posts is through index. Keeping sustainability in mind,
I only want to implement the routes that I intend to use. The reason for just using index
is because the only requirement is to list the news from HN using a link. Therefore,
I don't need to do other CRUD actions because this application is just to read data (other than favoriting).